+++ packages/aeria-sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # aeria-sdk
 
+## 0.0.25
+
+### Patch Changes
+
+- Fix context and SDK types
+- Updated dependencies
+  - @aeriajs/types@0.0.5
+  - @aeriajs/common@0.0.6
+
 ## 0.0.24
 
 ### Patch Changes