+++ packages/api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aeriajs/api
 
+## 0.0.13
+
+### Patch Changes
+
+- Fix context and SDK types
+- Updated dependencies
+  - @aeriajs/types@0.0.5
+  - @aeriajs/access-control@0.0.6
+  - @aeriajs/builtins@0.0.13
+  - @aeriajs/common@0.0.6
+  - @aeriajs/entrypoint@0.0.6
+  - @aeriajs/http@0.0.9
+  - @aeriajs/security@0.0.13
+  - @aeriajs/validation@0.0.9
+
 ## 0.0.12
 
 ### Patch Changes