+++ packages/common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aeriajs/common
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @aeriajs/types@0.0.5
+
 ## 0.0.5
 
 ### Patch Changes