+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aeriajs/types
 
+## 0.0.5
+
+### Patch Changes
+
+- Fix context and SDK types
+
 ## 0.0.4
 
 ### Patch Changes