+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aeriajs/validation
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @aeriajs/types@0.0.5
+  - @aeriajs/common@0.0.6
+
 ## 0.0.8
 
 ### Patch Changes